third_party/blink/web_tests/external/wpt/css/selectors/invalidation/has-in-ancestor-position-expected.txt

This is a testharness.js-based test.
Found 81 tests; 44 PASS, 37 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Initial color
PASS add .test to subject_ancestor
PASS remove .test from subject_ancestor
FAIL add .test to subject_parent assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from subject_parent
FAIL add .test to subject assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from subject
FAIL add .test to subject_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from subject_child
FAIL add .test to subject_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from subject_descendant
FAIL add .test to next_sibling assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove .test from next_sibling
FAIL add .test to next_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove .test from next_sibling_child
FAIL add .test to next_sibling_descendant assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove .test from next_sibling_descendant
PASS insert element div.test before subject_ancestor
PASS remove element div.test before subject_ancestor
FAIL insert element div.test before subject_parent assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before subject_parent
FAIL insert element div.test before subject assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before subject
FAIL insert element div.test before subject_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before subject_child
FAIL insert element div.test before subject_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before subject_descendant
FAIL insert element div.test before next_sibling assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove element div.test before next_sibling
FAIL insert element div.test before next_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove element div.test before next_sibling_child
FAIL insert element div.test before next_sibling_descendant assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove element div.test before next_sibling_descendant
FAIL insert element div.test after subject_ancestor assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove element div.test after subject_ancestor
FAIL insert element div.test after subject_parent assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove element div.test after subject_parent
FAIL insert element div.test after subject assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after subject
FAIL insert element div.test after subject_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after subject_child
FAIL insert element div.test after subject_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after subject_descendant
FAIL insert element div.test after next_sibling assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after next_sibling
FAIL insert element div.test after next_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove element div.test after next_sibling_child
FAIL insert element div.test after next_sibling_descendant assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove element div.test after next_sibling_descendant
PASS insert tree div>div.test before subject_ancestor
PASS remove tree div>div.test before subject_ancestor
FAIL insert tree div>div.test before subject_parent assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before subject_parent
FAIL insert tree div>div.test before subject assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before subject
FAIL insert tree div>div.test before subject_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before subject_child
FAIL insert tree div>div.test before subject_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before subject_descendant
FAIL insert tree div>div.test before next_sibling assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before next_sibling
FAIL insert tree div>div.test before next_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before next_sibling_child
FAIL insert tree div>div.test before next_sibling_descendant assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before next_sibling_descendant
FAIL insert tree div>div.test after subject_ancestor assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after subject_ancestor
FAIL insert tree div>div.test after subject_parent assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after subject_parent
FAIL insert tree div>div.test after subject assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after subject
FAIL insert tree div>div.test after subject_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after subject_child
FAIL insert tree div>div.test after subject_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after subject_descendant
FAIL insert tree div>div.test after next_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after next_sibling
FAIL insert tree div>div.test after next_sibling_child assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after next_sibling_child
FAIL insert tree div>div.test after next_sibling_descendant assert_equals: expected "rgb(255, 192, 203)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after next_sibling_descendant
Harness: the test ran to completion.
